web-port M2: clear all 63 WASM compile failures -> 743/743 TUs compile
Mechanical portability fixes so every engine TU compiles to a wasm32 object with em++ 3.1.69. Desktop i386
Linux build is preserved (verified: the unguarded changes still build with g++
-m32 -fms-extensions; web-only deltas are #ifdef __EMSCRIPTEN__-guarded).

Buckets:
* Pointer -> smaller-int casts (alignment-test idioms, packed-handle extraction):
  route through (uintptr_t) before the narrowing cast. ~40 sites across
  EffectsCore, gfx_d3d, clientscript, physics, sound, qcommon, universal, tl.
* operator new(unsigned int) -> operator new(size_t) (cscr_debugger.h).
* Bink/RAD: treat __EMSCRIPTEN__ as the Linux RAD target in radbase.h (no x86/MMX
  defines) so the platform-detect #error clears; Bink stays stubbed.
* va_list passed by non-const ref: declare real va_list locals, use va_end
  instead of `= 0` (ddl_api, r_warn, q_parse).
* noexcept-vs-musl: drop the exception-spec on the malloc/getenv redeclarations
  under __EMSCRIPTEN__ (corecrt_malloc.h, ShlObj.h).
* static-after-non-static: give des3_done external linkage (matches tomcrypt
  header; libtomcrypt des.c is not compiled here); drop the duplicate file-local
  static R_CullSphereDpvs (identical to the global in r_dpvs.cpp).
* placement new: #include <new> in phys_local.h.
* void <-> fn-ptr casts: explicit reinterpret_cast (rb_resource,
  com_expressions_eval, jdatasrc.c).
* speex_config_types.h (a generated header) absent: add an __EMSCRIPTEN__ branch
  in speex_types.h with the typedefs configure would emit.
* setjmp/longjmp: musl has no glibc __sigsetjmp/_longjmp. Bridge _setjmp/setjmp/
  longjmp over struct __jmp_buf_tag and widen g_script_error's jmp_buf backing
  to 64 ints (musl jmp_buf is 184 B vs i386's 64 B) so setjmp can't clobber
  neighbours.
* _mm_prefetch: include <xmmintrin.h> on wasm where the compat macro is
  suppressed (r_water, snd_dsp). _m_empty (MMX EMMS): no-op stub on wasm.
* ReadFileEx: declare LPOVERLAPPED_COMPLETION_ROUTINE and use it (Clang won't
  implicitly convert a function name to void*); fd_set via <sys/select.h>;
  GetProcessAffinityMask takes DWORD* (matches the call); const serverFilters[]
  value-initialized.

+++ src/EffectsCore/fx_marks.cpp
@@ -1975,7 +1975,7 @@ void __cdecl FX_EmitMarkTri(
     if ( marksSystem->hasCarryIndex )
     {
         pIndexa = (r_double_index_t *)&outSurf->indices[outSurf->indexCount - 1];
-        if ( ((unsigned __int8)pIndexa & 3) != 0
+        if ( ((unsigned __int8)(uintptr_t)pIndexa & 3) != 0
             && !Assert_MyHandler(
                         "C:\\projects_pc\\cod\\codsrc\\src\\EffectsCore\\fx_marks.cpp",
                         1684,
@@ -1994,7 +1994,7 @@ void __cdecl FX_EmitMarkTri(
     else
     {
         pIndex = (r_double_index_t *)&outSurf->indices[outSurf->indexCount];
-        if ( ((unsigned __int8)pIndex & 3) != 0
+        if ( ((unsigned __int8)(uintptr_t)pIndex & 3) != 0
             && !Assert_MyHandler(
                         "C:\\projects_pc\\cod\\codsrc\\src\\EffectsCore\\fx_marks.cpp",
                         1670,
