+++ src/database/db_file_load.cpp
@@ -541,7 +541,7 @@ char __cdecl DB_LoadXFile(
                 int allocType,
                 int flags)
 {
-    if ( ((unsigned __int8)buf & 3) != 0
+    if ( ((unsigned __int8)(uintptr_t)buf & 3) != 0
         && !Assert_MyHandler(
                     "C:\\projects_pc\\cod\\codsrc\\src\\database\\db_file_load.cpp",
                     1959,
