+++ src/gfx_d3d/r_sprite.cpp
@@ -351,7 +351,7 @@ void __cdecl R_GenerateQuadStampCodeMeshVertsArray(
         *(_QWORD *)prototype.v = *(_QWORD *)&prototypeVert.color.packed;
         *(_QWORD *)&prototype.unitVec[2].packed = __PAIR64__(v23.packed, v24.packed);
         verts = R_GetCodeMeshVerts(baseVertex);
-        if ( ((unsigned __int8)verts & 0x1F) != 0
+        if ( ((unsigned __int8)(uintptr_t)verts & 0x1F) != 0
             && !Assert_MyHandler(
                         "C:\\projects_pc\\cod\\codsrc\\src\\gfx_d3d\\r_sprite.cpp",
                         364,
