+++ tl/jobqueue/jobqueue_all.cpp
@@ -835,7 +835,7 @@ bool __cdecl jqPoll(jqBatchGroup *GroupID)
     p_group = GroupID;
     if (!GroupID)
         p_group = &jqPool.group;
-    if (((unsigned __int8)p_group & 7) != 0
+    if (((unsigned __int8)(uintptr_t)p_group & 7) != 0
         && _tlAssert(
             "c:\\projects_pc\\cod\\codsrc\\tl\\jobqueue\\jobqueue.cpp",
             390,
@@ -933,7 +933,7 @@ void __cdecl jqCheckDMALS(const void *addr)
 {
   if ( !addr )
     tlFatal("%s (LS) is NULL.", jqCheckContext);
-  if ( ((unsigned __int8)addr & 0xF) != 0 )
+  if ( ((unsigned __int8)(uintptr_t)addr & 0xF) != 0 )
     tlFatal("%s 0x%x (LS) not 16byte aligned.", jqCheckContext, addr);
   if ( (unsigned int)addr > 0x40000 )
     tlFatal("%s 0x%x (LS) is > 256k.", jqCheckContext, addr);
@@ -945,7 +945,7 @@ void __cdecl jqCheckDMAMain(const void *addr)
 {
   if ( !addr )
     tlFatal("%s (Main) is NULL.", jqCheckContext);
-  if ( ((unsigned __int8)addr & 0xF) != 0 )
+  if ( ((unsigned __int8)(uintptr_t)addr & 0xF) != 0 )
     tlFatal("%s 0x%x (Main) not 16byte aligned.", jqCheckContext, addr);
   if ( (unsigned int)addr < 0x40000 )
     tlFatal("%s 0x%x (Main) is < 256k.", jqCheckContext, addr);
