+++ tl/physics/rbc_def_vehicle.cpp
@@ -200,7 +200,7 @@ void rigid_body_constraint_wheel::get_wheel_collide_segment(
     //
     //v17[0] = a2;
     //v17[1] = p1a;
-    if (((unsigned __int8)p0 & 0xF) != 0
+    if (((unsigned __int8)(uintptr_t)p0 & 0xF) != 0
         && _tlAssert(
             "c:\\projects_pc\\cod\\codsrc\\tl\\physics\\include\\phys_math.h",
             444,
@@ -209,7 +209,7 @@ void rigid_body_constraint_wheel::get_wheel_collide_segment(
     {
         __debugbreak();
     }
-    if (((unsigned __int8)p1 & 0xF) != 0
+    if (((unsigned __int8)(uintptr_t)p1 & 0xF) != 0
         && _tlAssert(
             "c:\\projects_pc\\cod\\codsrc\\tl\\physics\\include\\phys_math.h",
             444,
@@ -452,7 +452,7 @@ void rigid_body_constraint_wheel::setup_constraint(pulse_sum_constraint_solver *
     this->m_ps_side_fric = 0;
     this->m_ps_fwd_fric = 0;
     ps_wheel = (pulse_sum_wheel *)this->m_wheel_flags;
-    if (((unsigned __int8)ps_wheel & 1) != 0)
+    if (((unsigned __int8)(uintptr_t)ps_wheel & 1) != 0)
     {
         b1 = this->b1;
         phys_multiply(&b1_suspension_dir, &this->b1->m_mat, &this->m_b1_suspension_dir_loc);
